Changed styles for home elements. Separated components to appropriate files.
